662 Strange behaviour of CUDS remove (DataspaceSession) (#668)
Allow the CUDS objects to be initialized with extra triples, instead of having to add them after already having created the object.

This solves #668 because previously, the class of the `Wrapper` CUDS objects was not defined on the server until after having spawned the CUDS object, because it is created empty and then filled with triples. This caused it to go to the added buffer, ignoring the changes introduced in #638, which solved the issue for local sessions, for which the CUDS objects are not created first empty and then filled with triples.