Support "Actions" in SimPhony
Adds support for registering methods as "actions" associated to ontology classes. For example, download and upload commands for files are now implemented in this way. But it would also be possible to write actions that multiply EMMO vectors, retrieve arrays from a URL, convert EMMO units, download a DCAT dataset, ...