Support "Operations" in SimPhoNy (#783)
Adds support for registering methods as "operations" associated to ontology classes. For example, download and upload commands for files are now implemented in this way. But it would also be possible to write operations that multiply EMMO vectors, retrieve arrays from a URL, convert EMMO units, download a DCAT dataset, ...

* Read operations from Python files in ~./simphony_osp/operations (also configurable via the environment variable `SIMPHONY_OPERATIONS_DIR`).

* Include the `Operations` abstract class in module `simphony_osp.develop`.

* Include the new `find_operations` function in the `simphony_osp.develop`. This function generates entry point declarations for operations automatically.

* Modify the file class so that it uses the Operations API.

* Modify the container class so that it uses the Operations API.